requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.6.0
attrs==23.2.0
bardapi==1.0.0
bcrypt==4.2.0
blinker==1.5
blurhash==1.1.4
boost==0.1
browser-cookie3==0.19.1
certifi==2022.12.7
cffi==1.16.0
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
cmake==3.26.3
colorama==0.4.6
comtypes==1.1.14
crispy-bootstrap5==0.7
decorator==5.1.1
dj-database-url==2.1.0
Django==4.2
django-crispy-forms==2.0
django-heroku==0.3.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
googletrans==3.1.0a0
greenlet==2.0.2
gunicorn==23.0.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.12.1
httpcore==1.0.5
httpx==0.27.0
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.2
lz4==4.3.3
MarkupSafe==2.1.1
Mastodon.py==1.8.1
mysql-connector==2.2.9
mysql-connector-python==9.0.0
mysqlclient==2.1.1
numpy==1.24.2
opencv-python==4.10.0.82
outcome==1.3.0.post0
packaging==24.1
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.3
psycopg2==2.9.7
PyAudio==0.2.13
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.0
pydantic_core==2.18.1
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.8.2
python-gemini-api==2.4.2
python-magic-bin==0.4.14
pyttsx3==2.90
pytube==15.0.0
pytube3==9.6.4
pytz==2022.7.1
pywin32==305
requests==2.28.1
rfc3986==1.5.0
selenium==4.19.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SpeechRecognition==3.9.0
SQLAlchemy==2.0.6
sqlparse==0.4.3
tk==0.1.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2023.3
urllib3==1.26.13
Werkzeug==2.2.2
whitenoise==6.4.0
win32-setctime==1.1.0
wsproto==1.2.0
WTForms==3.0.1
